requirements.txt
GitPython
radon 
coverage
openai
google-generativeai
anthropic
astor
astroid
langchain
pydantic
